Initial Commit | ROS Ad-Hoc UAV  Swarm
This commit includes the packages required to run the first stage of the dynamic ad-hoc UAV swarm. 

The files allow for any number of simulated UAVs running PX4 to be loaded into a gazebo world. The UAVs will position themselves into formation and follow the acting master in a square pattern.

The swarm nodes each communicate with one another. Each node in the network knows the position and velocity of every other node in the network. This allows for inter-swarm collision avoidance to be implemented. (which is next)